refactor: move impl snapshot5 from stream for impl_/stream
