+++ src/impl_/stream.rs
@@ -12,6 +12,7 @@ use crate::impl_::stream_sink::StreamSink;
 use crate::lambda2;
 use crate::IsLambda3;
 use crate::IsLambda4;
+use crate::IsLambda5;
 
 use parking_lot::Mutex;
 use parking_lot::RwLock;
@@ -312,6 +313,42 @@ impl<A: Send + 'static> Stream<A> {
             cb,
             lambda2(
                 move |a: &A, b: &B| f.call(a, b, &cc.sample(), &cd.sample()),
+                deps,
+            ),
+        )
+    }
+
+    pub fn snapshot5<B, C, D, E, F, FN>(
+        &self,
+        cb: &Cell<B>,
+        cc: &Cell<C>,
+        cd: &Cell<D>,
+        ce: &Cell<E>,
+        mut f: FN,
+    ) -> Stream<F>
+    where
+        B: Send + Clone + 'static,
+        C: Send + Clone + 'static,
+        D: Send + Clone + 'static,
+        E: Send + Clone + 'static,
+        F: Send + Clone + 'static,
+        FN: IsLambda5<A, B, C, D, E, F> + Send + Sync + 'static,
+    {
+        let mut deps = if let Some(deps2) = f.deps_op() {
+            deps2.clone()
+        } else {
+            Vec::new()
+        };
+        let cc = cc.clone();
+        let cd = cd.clone();
+        let ce = ce.clone();
+        deps.push(cc.to_dep());
+        deps.push(cd.to_dep());
+        deps.push(ce.to_dep());
+        self.snapshot(
+            cb,
+            lambda2(
+                move |a: &A, b: &B| f.call(a, b, &cc.sample(), &cd.sample(), &ce.sample()),
                 deps,
             ),
         )
